Move ROM selection into SELECT menu; restore X to random game
- Extract RomSelectionMenu from ISimpleGameListView.cpp anonymous
  namespace into GuiRomSelectionMenu (GuiSettings subclass) so it can
  be reused from GuiGamelistOptions
- Add "SELECT PREFERRED ROM" option in the SELECT menu (GuiGamelistOptions),
  shown only for GAME entries with at least one ROM in full UI mode
- Revert X button in ISimpleGameListView back to original behavior:
  jump to a random game in the current system
