+++ es-app/CMakeLists.txt
@@ -37,6 +37,7 @@ set(ES_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiCollectionSystemsOptions.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRandomCollectionOptions.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiInfoPopup.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomSelectionMenu.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomSelector.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomMetaDataEd.h
 
@@ -98,6 +99,7 @@ set(ES_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiCollectionSystemsOptions.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRandomCollectionOptions.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiInfoPopup.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomSelectionMenu.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomSelector.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/guis/GuiRomMetaDataEd.cpp
